refactor `target_locale` back to `locale`
I noticed that this was a bit messed up, I guess some of the AI tools had automatically changed some of the references from `target_locale` to `locale` but in an inconsistent manner. Since we are no longer importing the `locale` module here, we can safely use `locale` as an argument again.